Move coversion methods to CompanyHouse model
- Move mapToCompanyList & rsToCompanyList to the model
- Fix README formatting
- Fix error in CH.md
