+++ CH.md
@@ -42,7 +42,8 @@ SICCodeSicText3 string,
 SICCodeSicText4 string,
 LimitedPartnershipsNumGenPartners string,
 LimitedPartnershipsNumLimPartners string,
-URI,PreviousName1CONDATE string,
+URI string,
+PreviousName1CONDATE string,
 PreviousName1CompanyName string,
 PreviousName2CONDATE string,
 PreviousName2CompanyName string,
